Prettified kill claims on user profiles (#106)
* Added killer claims for self-reported deaths

If a victim reports their own death the server will auto-fill a
corresponding report from the killer. This is good because the killer
will want a claim to show up on their profile.

* Prettified kill claim reports on user pages

They're now grouped by day (and so also include a `timestamp` field),
with better descriptor strings.

* Added a statistics section to user pages

Fixes #46.