Fixed confirmation dialog styles
They weren't mobile friendly, and were glitchy on Chrome/Windows.

Previous style was to have two elements, an overlay and container, where
the overlay gave a translucent background and the container was
center-positioned via table displays with a white background. This is
risky structure and a bit complicated.

Now, the overlay contains the container, and they use flexbox for
centering.